+++ packages/datagrid/src/constants/index.ts
@@ -1,3 +1,2 @@
 export * from './datagrid.constants';
-export * from './column-definition.constants';
 export { AVRO_TYPES, LOGICAL_TYPES } from './avro-type.constant';