+++ packages/playground/src/app/components/ComponentFormSandbox.js
@@ -8,8 +8,8 @@ import ComponentForm from '@talend/react-containers/lib/ComponentForm';
 import SidePanel from '@talend/react-containers/lib/SidePanel';
 import HeaderBar from '@talend/react-containers/lib/HeaderBar';
 import { Alert } from '@talend/react-bootstrap';
-// eslint-disable-next-line @talend/import-depth
-import theme from '../example.scss';
+// test new behavior on non css module files
+import './ComponentFormSandbox.scss';
 
 const example = require('../../../mockBackend/mock/kit/example.json');
 const { isComponentFormDirty } = ComponentForm.selectors;
@@ -84,7 +84,7 @@ function ComponentFormSandBox({ dirty, dispatch }) {
 					window.location = '/';
 				}}
 			/>
-			<div id={theme.example}>
+			<div id="example">
 				{!hasAPI && (
 					<Alert>
 						You don t have backend API so we will use an ComponentForm as proxy to UIForm component