+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @talend/design-system
 
+## 5.2.0
+
+### Minor Changes
+
+- 38619790a: feat(design-system): add isFullWidth to StackHorizontal
+- fceb4c2f9: chore(design-system): ErrorState no longer WIP
+
+### Patch Changes
+
+- 9e653a037: fix(design-system): Link from ReactElement in ErrorState
+
+## 5.1.0
+
+### Minor Changes
+
+- 59509e0ba: feat(design-system): ErrorState (WIP)
+- dc361182d: Enabling `action` prop on EmptyState component size M.
+
+## 5.0.1
+
+### Patch Changes
+
+- e2174b30b: fix: scss filename now follow css module filename pattern
+- 6fd16be45: fix: use flex-start instead of start
+
 ## 5.0.0
 
 ### Major Changes